Rename file and all references to 'user'
